ROX-28979: address comments
