+++ ui/apps/platform/src/Containers/Clusters/ClustersTable.tsx
@@ -16,7 +16,7 @@ import { ClusterIdToRetentionInfo } from 'types/clusterService.proto';
 import { TableUIState } from 'utils/getTableUIState';
 
 import { formatCloudProvider } from './cluster.helpers';
-import ClustersStatusGrid from './ClustersStatusGrid';
+import ClusterStatusGrid from './ClusterStatusGrid';
 import { CertExpiryStatus } from './clusterTypes';
 import ClusterDeletion from './Components/ClusterDeletion';
 import ClusterNameWithTypeIcon from './Components/ClusterNameWithTypeIcon';
@@ -109,6 +109,11 @@ function ClustersTable({
                                 clusterInfo.status?.providerMetadata
                             );
                             const clusterId = clusterInfo.id;
+
+                            const isStatusUnavailable =
+                                !clusterInfo.healthStatus ||
+                                clusterInfo.healthStatus.overallHealthStatus === 'UNAVAILABLE';
+
                             return (
                                 <Tbody isExpanded={isRowExpanded(clusterId)} key={clusterInfo.id}>
                                     <Tr>
@@ -125,16 +130,23 @@ function ClustersTable({
                                         <Td dataLabel="Provider (Region)">{provider}</Td>
                                         <Td
                                             dataLabel="Cluster Status"
-                                            compoundExpand={{
-                                                isExpanded: isCellExpanded(
-                                                    clusterId,
-                                                    EXPANDABLE_COLUMN.STATUS
-                                                ),
-                                                onToggle: () =>
-                                                    toggle(clusterId, EXPANDABLE_COLUMN.STATUS),
-                                                rowIndex,
-                                                columnIndex: 3,
-                                            }}
+                                            compoundExpand={
+                                                !isStatusUnavailable
+                                                    ? {
+                                                          isExpanded: isCellExpanded(
+                                                              clusterId,
+                                                              EXPANDABLE_COLUMN.STATUS
+                                                          ),
+                                                          onToggle: () =>
+                                                              toggle(
+                                                                  clusterId,
+                                                                  EXPANDABLE_COLUMN.STATUS
+                                                              ),
+                                                          rowIndex,
+                                                          columnIndex: 3,
+                                                      }
+                                                    : undefined
+                                            }
                                         >
                                             <ClusterStatus
                                                 healthStatus={clusterInfo?.healthStatus}
@@ -194,17 +206,19 @@ function ClustersTable({
                                             />
                                         </Td>
                                     </Tr>
-                                    {isCellExpanded(clusterId, EXPANDABLE_COLUMN.STATUS) && (
-                                        <Tr isExpanded>
-                                            <Td colSpan={colSpan}>
-                                                <ExpandableRowContent>
-                                                    <ClustersStatusGrid
-                                                        healthStatus={clusterInfo.healthStatus}
-                                                    />
-                                                </ExpandableRowContent>
-                                            </Td>
-                                        </Tr>
-                                    )}
+                                    {clusterInfo.healthStatus &&
+                                        isCellExpanded(clusterId, EXPANDABLE_COLUMN.STATUS) && (
+                                            <Tr isExpanded>
+                                                <Td colSpan={colSpan}>
+                                                    <ExpandableRowContent>
+                                                        <ClusterStatusGrid
+                                                            healthStatus={clusterInfo.healthStatus}
+                                                        />
+                                                    </ExpandableRowContent>
+                                                </Td>
+                                            </Tr>
+                                        )}
+
                                     {isCellExpanded(clusterId, EXPANDABLE_COLUMN.SENSOR) && (
                                         <Tr isExpanded>
                                             <Td colSpan={colSpan}>
