+++ ui/apps/platform/src/Containers/Clusters/cluster.constants.ts
@@ -1,0 +1,7 @@
+export const healthStatusLabels = Object.freeze({
+    UNINITIALIZED: 'Uninitialized',
+    UNAVAILABLE: 'Unavailable',
+    UNHEALTHY: 'Unhealthy',
+    DEGRADED: 'Degraded',
+    HEALTHY: 'Healthy',
+});
